refactor: split AdminGeneralSettingsForm into focused sub-components
